Moved model to separate file
